Aufgabe-2 weiter

+++ RMI/Programme/Aufgabe-2/src/Quadrat.java
@@ -7,6 +7,6 @@ import java.rmi.RemoteException;
 
 public interface  Quadrat  extends Remote {             //Interface des Remote Objekts
 
-         double quadrat(double i) throws RemoteException;
+         long quadrat(long i) throws RemoteException;
 
 }
